Add files via upload
Please see all the cata comments on how each line of code works. If you have any doubt on the functions, look at the documentation of the library used!! There will also just be errors since this hasn't been finally tested with the EMG so feel free to slap this into chatgpt or deepseek or whatever else to fix stuff that isn't working.